kraken/server/requirements.txt
billiard==3.6.1.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.3.2
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
kombu==4.6.5
MarkupSafe==1.1.0
more-itertools==7.2.0
openapi-spec-validator==0.2.4
pathlib==1.0.1
psycopg2==2.8.3
Pygments==2.4.2
pytz==2019.2
PyYAML==4.2b4
redis==3.3.8
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
swagger-ui-bundle==0.0.3
urllib3==1.24.1
vine==1.3.0
Werkzeug==0.14.1
zipp==0.6.0
apscheduler==3.6.1
json-rpc==1.12.1
jsonrpcclient[requests]==3.3.4
pytimeparse==1.1.8
python-dateutil==2.8.0
